Refactor Order Service  Remove Unused Code, Rename Order Entity Variables, and Update Migrations
