+++ src/entities/setting.entity.ts
